feature: Dropped command differentiators
Basically, silent and react functions were supposed to have much larger
differences. On review, however, it became clear that there was little to
differentiate in the base code save setting up the response functionality
and timeouts - all easily stuffed into an if statement.

For the sake of clarity I've toned silentFncs down to a setting in the
command options, with teh default being react functions.

Now there is simply "commands"